Mapped job duration to note frequency.
dj-mqtt.py:
Created a scale of notes based on natural harmonics. Log is then taken of the duration of a job and is assigned to a note in the scale based on the result. Can set whether short jobs are high notes or low. High is preferred by me so far.

duration-graph.py:
Just set it to map a histogram of log(duration) not duration.

render2.rb:
This is the sonic pi file to go with the new version of dj-mqtt.py. Render.rb kept as an example for myself as I'm unfamiliar with sonic pi.