Mapped job site to pan
dj-mqtt.py:
Added hashToPan().
This takes a string, hashes it and converts that hash to a float -1 to 1 representing pan.
Added lines to split job site from message, calculate pan and send it in osc message.

hashTest.py:
Code to test functionality of hashToPan() and get a feel for the stereo distribution it creates.

render2.rb:
Added cutoff variable and also added pan variable which takes value from osc message.